requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0

python-dotenv==1.0.1

google-generativeai==0.8.3
openai==1.52.0

pillow==10.4.0
numpy==2.1.0

python-multipart==0.0.9

firebase-admin==6.6.0